+++ packages/core/src/core/config/config-mock.spec.ts
