+++ packages/core/src/core/config/config-mock.ts
