+++ packages/core/src/core/config/index.ts
@@ -1,0 +1,2 @@
+export { Config, ConfigNotFoundError, ConfigTypeError } from './config';
+export { ConfigMock } from './config-mock';
