+++ packages/core/src/core/index.ts
@@ -3,6 +3,5 @@ export { createController } from './controllers';
 export * from './http';
 export * from './hooks';
 export * from './routes';
-export { ConfigMock } from './config-mock';
-export { Config, ConfigNotFoundError, ConfigTypeError } from './config';
+export { Config, ConfigMock, ConfigNotFoundError, ConfigTypeError } from './config';
 export { createService, dependency, ServiceManager } from './service-manager';
